% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichplot.r
\name{enrichplot}
\alias{enrichplot}
\title{enrichplot function is used to draw the enrichment plot.}
\usage{
enrichplot(file, type)
}
\arguments{
\item{file}{pathway file, the format is xls}

\item{type}{the type of plot, dot,line,barcode}
}
\value{
a plot, the format is pdf
}
\description{
enrichplot function is used to draw the enrichment plot.
}
\examples{
path="D:\\\\code\\\\r\\\\FigureYa13GSEA_Java_update"
file <- mergefile(path)
enrichplot <-enrichplot(file,type="dot")
}
