% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergefile.r
\name{mergefile}
\alias{mergefile}
\title{mergefile function is used to merge all xls files in the working path.}
\usage{
mergefile(path)
}
\arguments{
\item{path}{the parameter is the working path and should be a string}
}
\value{
a data.frame.
}
\description{
mergefile function is used to merge all xls files in the working path.
}
\examples{
path="D:\\\\code\\\\r\\\\FigureYa13GSEA_Java_update"
result <- mergefile(path)
}
